+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Exceptions thrown in error handlers are ignored but logged (#3559).
 * Removed unused public constants in `SyncConfiguration` (#4047).
 * Fixed bug, preventing Sync client to renew the access token (#4038) (#4039).
+* Now `SyncUser.logout()` properly revoke tokens (#3639).
 
 ### Bug fixes